Fixed a bug preventing dir and symlink deletion
Also fixed a bug causing a single path to be printed repeatedly
(once for every item in `deferred_items`) at the end of execution.

`deferred_items` was a list of two-tuples; the loop iterating over it
assumed it was a list of paths. Also, `action` never got set in the loop;
it was probably retaining a value of Action.skip or Action.unlink from earlier.
Because of that, neither os.rmdir() nor os.unlink() was actually called,
resulting in unconditional success and thus unconditional printing of a second
leaked variable on each iteration! What a mess.

Anyway, both bugs should be fixed now.