Started saving and restoring directory timestamps
We now save the modification timestamps from directories we modify (i.e.
unlink items from) before the first time we modify them, and restore them
when we're done (after handling deferrals).

(Access timestamps are actually reset, too, but because a directory's timestamps
are saved after we've recursed into it (and updated its timestamp), not much
of a difference is made there. In order to handle atimes as mtimes are,
we would have to either a) store timestamps for /all/ directories in memory
(certainly possible, but costly for large directory hierarchies), or
b) gain foreknowledge of which directories will have items unlinked from them.
And because os.utime() doesn't seem to have a way to /not/ change access times,
in order to (/effectively/) not set them back, we would have to call os.stat()
right before os.utime() (which I'm intuitively inclined against doing, since
we would have already called it for the same directory earlier).)